add abstract class to capture training utilities to be shared across models
