+++ sup3r/models/__init__.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 """Sup3r Model Software"""
-from .generator import Sup3rGen
 from .base import Sup3rGan
 from .wind import WindGan
 from .solar_cc import SolarCC
